pyproject.toml
[tool.poetry]
name = "currencies"
version = "0.1.0"
description = ""
authors = ["resistansepower <resistansepower@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
drf-spectacular = "^0.26.5"
drf-yasg = "^1.21.7"
requests = "^2.31.0"
celery = "^5.3.4"
psycopg2 = "^2.9.7"
django-environ = "^0.11.2"
django-celery-beat = "^2.5.0"
redis = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
